Refactored serializers and removed most nested serializers
